pyproject.toml
[tool.poetry]
name = "volmir"
version = "0.1.0"
description = ""
authors = ["Jere Perisic <perisic.j@northeastern.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.2.4"
djangorestframework = "^3.16.0"
django-cors-headers = "^4.7.0"
pydantic = "^2.11.7"
pillow = "^11.3.0"
pyjwt = "^2.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
